+++ software/main_controller/components/hal/src/arduino/time.cpp
@@ -10,13 +10,13 @@
 #include <Arduino.h>
 #include "hal/time.hpp"
 
-void hal::Delay(uint64_t ms)
+void hal::Delay(uint32_t ms)
 {
     delay(ms);
 }
 
 
-uint64_t hal::GetUptime_ms()
+uint32_t hal::GetUptime_ms()
 {
-    return static_cast<uint64_t>(millis());
+    return millis();
 }
